add custom exception handle

+++ HR.LeaveManagement/HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -25,6 +27,12 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Command
 		public async Task Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
 		{
 			var leaveRequest = await leaveRequestRepository.GetAsync(request.Id);
+			
+			if (leaveRequest == null)
+			{
+				throw new NotFoundException(nameof(LeaveRequest), request.Id);
+			}
+
 			await leaveRequestRepository.DeleteAsync(leaveRequest);
 		}
 	}
